backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
PyJWT==2.9.0
bcrypt==4.2.0
python-multipart==0.0.12
pydantic==2.9.2
